Solution
I don't like this problem. I was able to get the right solution after figuring out how class template specialization works on learncpp. The issue I have with it is that it makes no sense. The whole point of templates is to avoid writing out lots of code, but this problem makes you do just that. This is  essentially an inheritance problem which forces you to use template specialization instead.